ai-service/pyproject.toml
[project]
name = "affiliate-link"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.120.3",
    "httpx[http2]>=0.28.1",
    "langchain-core>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
    "uvicorn>=0.38.0",
]
